Oauth refactoring : Change package name
