+++ src/eclipse/plugins/com.ibm.sbt.core/META-INF/MANIFEST.MF
@@ -18,8 +18,8 @@ Export-Package: com.ibm.sbt.core.configuration,
  com.ibm.sbt.security.authentication,
  com.ibm.sbt.security.authentication.oauth,
  com.ibm.sbt.security.authentication.oauth.consumer,
- com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.servlet,
- com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.util,
+ com.ibm.sbt.security.authentication.oauth.consumer.servlet,
+ com.ibm.sbt.security.authentication.oauth.consumer.util,
  com.ibm.sbt.security.authentication.oauth.consumer.store,
  com.ibm.sbt.security.authentication.password,
  com.ibm.sbt.security.authentication.password.consumer,
