+++ src/eclipse/plugins/com.ibm.sbt.core/src/com/ibm/sbt/security/authentication/oauth/consumer/OAProvider.java
@@ -31,8 +31,8 @@ import com.ibm.commons.util.profiler.ProfilerAggregator;
 import com.ibm.commons.util.profiler.ProfilerType;
 import com.ibm.sbt.core.configuration.Configuration;
 import com.ibm.sbt.security.authentication.oauth.OAuthException;
-import com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.servlet.OACallback;
-import com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.util.OADance;
+import com.ibm.sbt.security.authentication.oauth.consumer.servlet.OACallback;
+import com.ibm.sbt.security.authentication.oauth.consumer.util.OADance;
 import com.ibm.sbt.security.authentication.oauth.consumer.store.OATokenStoreFactory;
 import com.ibm.sbt.security.authentication.oauth.consumer.store.TokenStore;
 import com.ibm.sbt.service.util.ServiceUtil;
