+++ src/eclipse/plugins/com.ibm.sbt.core/src/com/ibm/sbt/security/authentication/oauth/consumer/servlet/OA2Callback.java
@@ -13,7 +13,7 @@
  * implied. See the License for the specific language governing 
  * permissions and limitations under the License.
  */
-package com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.servlet;
+package com.ibm.sbt.security.authentication.oauth.consumer.servlet;
 
 import java.io.IOException;
 import java.text.MessageFormat;
