+++ src/eclipse/plugins/com.ibm.sbt.core/src/com/ibm/sbt/security/authentication/oauth/consumer/util/OADance.java
@@ -13,7 +13,7 @@
  * implied. See the License for the specific language governing 
  * permissions and limitations under the License.
  */
-package com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.util;
+package com.ibm.sbt.security.authentication.oauth.consumer.util;
 
 import java.io.IOException;
 import java.io.Serializable;
