+++ src/eclipse/plugins/com.ibm.sbt.core/src/com/ibm/sbt/service/core/servlet/ServiceServlet.java
@@ -18,9 +18,9 @@ package com.ibm.sbt.service.core.servlet;
 import com.ibm.commons.runtime.RuntimeConstants;
 import com.ibm.commons.runtime.servlet.ServletDispatcher;
 import com.ibm.commons.runtime.servlet.ServletFactory;
-import com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.servlet.OA2Callback;
-import com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.servlet.OACallback;
-import com.ibm.sbt.security.authentication.oauth.consumer.oauth_10a.servlet.OAClientAuthentication;
+import com.ibm.sbt.security.authentication.oauth.consumer.servlet.OA2Callback;
+import com.ibm.sbt.security.authentication.oauth.consumer.servlet.OACallback;
+import com.ibm.sbt.security.authentication.oauth.consumer.servlet.OAClientAuthentication;
 import com.ibm.sbt.service.core.handlers.AuthenticationHandler;
 import com.ibm.sbt.service.core.handlers.BasicAuthCredsHandler;
 import com.ibm.sbt.service.core.handlers.EmailHandler;
